Move Cosmos integration tests into tests subdirectory
