+++ crates/cosmos/cosmos-integration-tests/src/lib.rs
@@ -5,6 +5,3 @@ extern crate alloc;
 
 pub mod contexts;
 pub mod impls;
-
-#[cfg(test)]
-pub mod tests;
