--- crates/cosmos/cosmos-integration-tests/src/tests/mod.rs
@@ -1,2 +1,0 @@
-pub mod celestia;
-pub mod cosmos;
