+++ crates/cosmos/cosmos-integration-tests/tests/celestia.rs
@@ -1,3 +1,5 @@
+#![recursion_limit = "256"]
+
 use core::time::Duration;
 use std::sync::Arc;
 
@@ -9,15 +11,15 @@ use hermes_relayer_runtime::types::runtime::HermesRuntime;
 use hermes_test_components::setup::traits::run_test::CanRunTest;
 use ibc_relayer::chain::client::ClientSettings;
 use ibc_relayer::chain::cosmos::client::Settings;
+use ibc_relayer::config::compat_mode::CompatMode;
 use ibc_relayer_types::core::ics02_client::trust_threshold::TrustThreshold;
 use ibc_relayer_types::core::ics24_host::identifier::PortId;
 use tokio::runtime::Builder;
-
-use crate::contexts::binary_channel::setup::CosmosBinaryChannelSetup;
-use crate::contexts::bootstrap::CosmosBootstrap;
+use hermes_cosmos_integration_tests::contexts::binary_channel::setup::CosmosBinaryChannelSetup;
+use hermes_cosmos_integration_tests::contexts::bootstrap::CosmosBootstrap;
 
 #[test]
-fn cosmos_integration_tests() -> Result<(), Error> {
+fn celestia_integration_tests() -> Result<(), Error> {
     let _ = stable_eyre::install();
 
     let tokio_runtime = Arc::new(Builder::new_multi_thread().enable_all().build()?);
@@ -26,8 +28,21 @@ fn cosmos_integration_tests() -> Result<(), Error> {
 
     let builder = Arc::new(CosmosBuilder::new_with_default(runtime.clone()));
 
-    // TODO: load parameters from environment variables
-    let bootstrap = Arc::new(CosmosBootstrap {
+    let celestia_bootstrap = Arc::new(CosmosBootstrap {
+        runtime: runtime.clone(),
+        builder: builder.clone(),
+        should_randomize_identifiers: true,
+        test_dir: "./test-data".into(),
+        chain_command_path: "celestia-appd".into(),
+        account_prefix: "celestia".into(),
+        compat_mode: Some(CompatMode::V0_34),
+        staking_denom: Denom::base("utia"),
+        transfer_denom: Denom::base("coin"),
+        genesis_config_modifier: Box::new(|_| Ok(())),
+        comet_config_modifier: Box::new(|_| Ok(())),
+    });
+
+    let cosmos_bootstrap = Arc::new(CosmosBootstrap {
         runtime,
         builder,
         should_randomize_identifiers: true,
@@ -43,20 +58,19 @@ fn cosmos_integration_tests() -> Result<(), Error> {
 
     let create_client_settings = ClientSettings::Tendermint(Settings {
         max_clock_drift: Duration::from_secs(40),
-        trusting_period: None,
+        trusting_period: Some(Duration::from_secs(60 * 60)),
         trust_threshold: TrustThreshold::ONE_THIRD,
     });
 
     let setup = CosmosBinaryChannelSetup {
-        bootstrap_a: bootstrap.clone(),
-        bootstrap_b: bootstrap,
+        bootstrap_a: celestia_bootstrap,
+        bootstrap_b: cosmos_bootstrap,
         create_client_settings,
         init_connection_options: Default::default(),
         init_channel_options: Default::default(),
         port_id: PortId::transfer(),
     };
 
-    // TODO: Use a test suite entry point for running multiple tests
     tokio_runtime.block_on(async move {
         setup.run_test(&TestIbcTransfer).await?;
